+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - SC2262/SC2263: Warn about aliases declared and used in the same parsing unit
 - SC2264: Warn about wrapper functions that blatantly recurse
 - SC2265/SC2266: Warn when using & or | with test statements
+- SC2267: Warn when using xargs -i instead of -I
 
 ### Fixed
 - SC1072/SC1073 now respond to disable annotations, though ignoring parse errors